requirements.txt
Django==1.6.1
djangorestframework==2.3.12
MySQL-python==1.2.3
bae_log